Created Dynamic Web Project "capstone"
This is a new project folder on eclipse with none of the stuff we've
contributed on it. It will eventually hold all the contributions in the
WEB-INF and the src folders. (I will be copy and pasting them manually
from the project in the "JSakian/capstone" repo into this project
"capstone" which is in Jsakian/gitrdone)