Corrects logic for Events variable
a single event is comprised of a name, a date, a description, and those
who will be attending.

The way the variable was written prior, it was suggesting the Event
class was only the event's name.

Since this is a Person class, and it can hold the events each individual
will attend, the eventName variable was changed to attendingEvents, the
type of the variable was changed to an array list of type Event, and the
getters and setters were changed to reflect all changes.