Removes debugging console prints & fixes volunteering value issue
The volunteering value issue was:
after setting the volunteering boolean to true in the volunteerPerson
object, the value would reset to false. This was because the boolean
volunteering wasn't considered as a backing object in the form so when
the formSubmission method attempted to use the value to determine a
command, the value would reset bc it wasn't supported as true on the
form. It has been corrected by changing the value of the boolean to true
when the form has been successfully submitted and the next method is
invoked.