Changes the "attendingEvents" variable to "event"
The fact that we could have one Person attending multiple events was the
reason this was an ArrayList. It was later discussed that the preferred
way of keeping track of attendees and volunteers would be through one
event. This way, the comments entered in the form by the user were
always specific to that event which it was related to.